Signup js, css분리 및 app.js 업데이트
