+++ frontend/src/App.js
@@ -3,7 +3,7 @@ import logo from './logo.svg';
 import './App.css';
 import React from 'react';
 
-import Signup from './pages/performer/Signup';
+import Signup from './pages/performer/Signup/Signup';
 
 function App() {
     return (
